make default menu bar include 'enabled' + add menu bar to font info and application settings
