+++ src/fontra/views/applicationsettings/applicationsettings.html
@@ -62,6 +62,7 @@
   </head>
   <body>
     <modal-dialog></modal-dialog>
+    <div class="top-bar-container cleanable-overlay"></div>
     <div class="main-container">
       <div id="header-container"></div>
       <div id="panel-container"></div>
