+++ src/fontra/views/fontinfo/fontinfo.html
@@ -60,6 +60,7 @@
   </head>
   <body>
     <modal-dialog></modal-dialog>
+    <div class="top-bar-container cleanable-overlay"></div>
     <div class="main-container">
       <div id="header-container"></div>
       <div id="panel-container"></div>
